refactor markdown sync logic
